Remove alb; use production public alb
**Why:**
We want to use an already existing alb instead of paying for one. This
should allow us to route to the listener directly for as many services
as we want. We may want to split this off for unrelated services since
there is a pretty strict inbound rule for the public albs

[ch24814] https://app.clubhouse.io/itprotv/story/24814